E2E Libification rebased (3).
